Fixing the build.
By bumping Shiro to 1.2.1 that contains fix for SHIRO-344,
build failed and revealed the fact that NexusHttpAuthenticationFilter
was successfully handling the anon-login-with-expired-session
case due to a bug.

As we use DelegatingSubject -- that is affected by this fix,
the method subject.getPrincipal() should not be used once we
know session is expired (we are handling the UnknownSessionException
exception!)

Also, as this happens in logging, and the intention seems to be
to log the affected "anon" username, the change is simply
to log the username actually used in log-in process.